Cambios e implementacion ModelMapper a UserController y UserServices

+++ src/main/java/com/app/clubnautico/services/UserServices.java
@@ -3,6 +3,7 @@ package com.app.clubnautico.services;
 import java.util.ArrayList;
 import java.util.Optional;
 
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -16,32 +17,31 @@ public class UserServices {
 	@Autowired
 	UserRepository userRepository; // Inyección de instancia del repositorio de UserRepository para interactuar con la base de datos
 
+	@Autowired
+	private ModelMapper modelMapper;
+	
 	public ArrayList<UserModel> getUsers() {
 		// Retorna una lista de todos los usuarios almacenados en la base de datos
 		return (ArrayList<UserModel>) userRepository.findAll();
 	}
 
-	public UserModel saveUser(UsuarioDTO userDTO) {
-		// Crea una instancia de UserModel y asigna los valores del DTO de usuario al modelo de usuario
-		UserModel user = new UserModel();
-		user.setDni(userDTO.getDni());
-		user.setNombre(userDTO.getNombre());
-		// Asigna otros campos
-		// Guarda el usuario en la base de datos y lo devuelve
-		return userRepository.save(user);
+	public UsuarioDTO saveUser(UsuarioDTO userDTO) {
+		//Convertimos el DTO a entidad para que interactue con la base de datos.
+		UserModel user = modelMapper.map(userDTO, UserModel.class);
+		//Guardamos la entidad en la base de datos.
+		UserModel usuarioGuardado = userRepository.save(user);
+		//Convertimos la entidad a DTO para devolver la respuesta
+		UsuarioDTO usuarioGuardadoDTO = modelMapper.map(usuarioGuardado, UsuarioDTO.class);
+		
+		return usuarioGuardadoDTO;
 	}
 
 	public Optional<UsuarioDTO> getUserById(Long id) {
 		// Busca un usuario por ID en la base de datos
 		Optional<UserModel> userOptional = userRepository.findById(id);
 		if (userOptional.isPresent()) {
-			// Si el usuario existe, crea un DTO de usuario y asigna los valores del usuario al DTO
-			UserModel user = userOptional.get();
-			UsuarioDTO userDTO = new UsuarioDTO();
-			//Aqui podemos definir lo que queremos que nos devuelva la aplicacion como respuesta, en este caso solo devuelve DNI y nombre aunque hayan mas datos.
-			userDTO.setDni(user.getDni());
-			userDTO.setNombre(user.getNombre());
-			
+			// Si el usuario existe, convierte la entidad a DTO usando modelMapper
+			UsuarioDTO userDTO = modelMapper.map(userOptional.get(), UsuarioDTO.class);
 			// Devuelve un Optional que contiene el DTO de usuario
 			return Optional.of(userDTO);
 		} else {
@@ -50,23 +50,35 @@ public class UserServices {
 		}
 	}
 
-	public UserModel UpdateUserById(UserModel request, Long id) {
-		// Busca el usuario por ID
-		UserModel user = userRepository.findById(id).get();
-		// Actualiza los campos del usuario con los valores del modelo de usuario recibido
-		user.setDni(request.getDni());
-		user.setDireccion(request.getDireccion());
-		user.setEmail(request.getEmail());
-		// Retorna el usuario actualizado
-		return user;
+	public Optional<UsuarioDTO> UpdateUserById(UsuarioDTO userDTO, Long id) {
+		// Buscamos el usuario por ID
+		Optional<UserModel> userOptional = userRepository.findById(id);
+		// Comprobamos si el usuario existe
+		if (userOptional.isPresent()) {
+			//Si existe el usuario, actualizamos los campos con los valores que tiene el DTO.
+			UserModel user = userOptional.get();
+			modelMapper.map(userDTO, user);
+			//Guardamos los cambios en la base de datos
+			UserModel updateUser = userRepository.save(user);
+			//Convertimos la entidad actualizada a DTO
+			UsuarioDTO updateUserDTO = modelMapper.map(updateUser, UsuarioDTO.class);
+			//Devolvemos el usuario actualizado en formato DTO
+			return Optional.of(updateUserDTO);
+		} else {
+			return Optional.empty();
+		}
 	}
 
 	public Boolean deleteUser(Long id) {
 		try {
-			// Elimina el usuario de la base de datos por ID
-			userRepository.deleteById(id);
-			// Retorna true si se elimina correctamente
-			return true;
+			//Verificamos si el usuario existe
+			Optional<UserModel> userOptional = userRepository.findById(id);
+			if (userOptional.isPresent()) {
+				userRepository.deleteById(id);
+				return true;
+			} else {
+				return false;
+			}
 		} catch (Exception e) {
 			// Retorna false si ocurre un error al eliminar el usuario
 			return false;
